Move `testData` to `FakeLms.fetchSemesters`
