+++ model/lms.js
@@ -14,7 +14,7 @@ class Lms {
 	/**
 	 * @param {string|null} bookmark - can be null
 	 * @param {string|null} search - can be null
-	 * @returns {{PagingInfo:{Bookmark: string, HasMoreItems: boolean}, Items: {orgUnitId: number, orgUnitName: string}[]]}
+	 * @returns {{PagingInfo:{Bookmark: string, HasMoreItems: boolean}, Items: {orgUnitId: number, orgUnitName: string}[]}}
 	 */
 	async fetchSemesters(bookmark, search) {
 		const url = new URL('/d2l/api/ap/unstable/insights/data/semesters', location.href);
